Analog abstraction

+++ examples/AWS_IoT_Core/AWS_IoT_Core.ino
@@ -99,7 +99,7 @@ const int8_t modemLEDPin =
     redLED;  // MCU pin connected an LED to show modem status
 
 // Network connection information
-const char* wifiId = "YourWiFiSSID";  // The WiFi access point
+const char* wifiId  = "YourWiFiSSID";      // The WiFi access point
 const char* wifiPwd = "YourWiFiPassword";  // The WiFi password
 
 // Create the loggerModem object
@@ -145,8 +145,8 @@ const int8_t alsData = A8;  // The ALS PT-19 data pin
 #else
 const int8_t alsData = A4;  // The ALS PT-19 data pin
 #endif
-const int8_t  alsSupply     = 3.3;  // The ALS PT-19 supply power voltage
-const int8_t  alsResistance = 10;   // The ALS PT-19 loading resistance (in kΩ)
+const float   alsSupply     = 3.3f;  // The ALS PT-19 supply power voltage
+const int8_t  alsResistance = 10;    // The ALS PT-19 loading resistance (in kΩ)
 const uint8_t alsNumberReadings = 10;
 
 // Create a Everlight ALS-PT19 sensor object
